add watermark tracer
adds a autotracer that steps through a program.
trace <address>
When the PC reaches <address> or higher the marker gets active. As soon as PC drops below the marker again,
the tracer stops. It prints the PC register on all steps.
This is usefull if crashes occure that result in a reboot. For example 'trace 0x801a' is a good marker for the
eZ430 Chronos Watch. If due some programming error, the watch reboots, the tracer stops and code parts causing the
reboot can be identified.